trouble-api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
anyio==3.3.4
    # via starlette
asgiref==3.4.1
    # via uvicorn
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via
    #   nltk
    #   sacremoses
    #   uvicorn
dataclasses-json==0.5.6
    # via -r requirements.in
fastapi==0.70.0
    # via -r requirements.in
filelock==3.3.2
    # via
    #   huggingface-hub
    #   transformers
googleapis-common-protos==1.53.0
    # via vald-client-python
grpcio==1.41.1
    # via vald-client-python
h11==0.12.0
    # via uvicorn
huggingface-hub==0.1.2
    # via
    #   sentence-transformers
    #   transformers
idna==3.3
    # via
    #   anyio
    #   requests
joblib==1.1.0
    # via
    #   nltk
    #   sacremoses
    #   scikit-learn
marshmallow==3.14.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
mypy-extensions==0.4.3
    # via typing-inspect
nltk==3.6.5
    # via sentence-transformers
numpy==1.21.4
    # via
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
packaging==21.2
    # via
    #   huggingface-hub
    #   transformers
pillow==8.4.0
    # via torchvision
protobuf==3.19.1
    # via
    #   googleapis-common-protos
    #   vald-client-python
pydantic==1.8.2
    # via fastapi
pyparsing==2.4.7
    # via packaging
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2021.11.10
    # via
    #   nltk
    #   sacremoses
    #   transformers
requests==2.26.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
sacremoses==0.0.46
    # via transformers
scikit-learn==1.0.1
    # via sentence-transformers
scipy==1.7.2
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.1.0
    # via -r requirements.in
sentencepiece==0.1.96
    # via sentence-transformers
six==1.16.0
    # via
    #   grpcio
    #   sacremoses
sniffio==1.2.0
    # via anyio
starlette==0.16.0
    # via fastapi
threadpoolctl==3.0.0
    # via scikit-learn
tokenizers==0.10.3
    # via
    #   sentence-transformers
    #   transformers
torch==1.10.0
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.11.1
    # via sentence-transformers
tqdm==4.62.3
    # via
    #   huggingface-hub
    #   nltk
    #   sacremoses
    #   sentence-transformers
    #   transformers
transformers==4.12.2
    # via
    #   -r requirements.in
    #   sentence-transformers
typing-extensions==3.10.0.2
    # via
    #   huggingface-hub
    #   pydantic
    #   torch
    #   typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.7
    # via requests
uvicorn==0.15.0
    # via -r requirements.in
vald-client-python==1.3.1
    # via -r requirements.in
